move log_shas and spoonfeed_log_shas to git_operations
